fix:fix test unit compilation errors (#3442)
